test: align github client tests with repo conventions
